+++ src/main.js
@@ -1,5 +1,4 @@
 import task from './data.js';
-// import generateCard from './make-task.js';
 import {filtersNames, generateFilter} from './make-filter.js';
 import Task from './task.js';
 import TaskEdit from './task-edit.js';
@@ -21,8 +20,9 @@ const renderFilters = () => {
 const renderTasks = (num) => {
   const fragment = document.createDocumentFragment();
   for (let i = 0; i < num; i++) {
-    const taskComponent = new Task(task());
-    const editTaskComponent = new TaskEdit(task());
+    const tasks = task();
+    const taskComponent = new Task(tasks);
+    const editTaskComponent = new TaskEdit(tasks);
 
     taskComponent.onEdit = () => {
       editTaskComponent.render();
